requirements.txt
qiskit
azure-quantum[qiskit]
numpy
pygame
